# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-isfinite",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — is `value` finite?",
        "description": "JSON numbers cannot encode infinities or NaN, so any value `srvcs-isnumber`\naccepts as a number is finite by construction. This service is therefore\nnear-degenerate: it delegates \"is this a number\" to `srvcs-isnumber` and\nreports that verdict as its `result`. If the dependency is unreachable, it\nreports itself degraded rather than guessing.",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredicateResponse"
                }
              }
            }
          },
          "422": {
            "description": "value is not a valid number (forwarded)"
          },
          "500": {
            "description": "an upstream dependency returned an unexpected response"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EvalRequest": {
        "type": "object",
        "required": [
          "value"
        ],
        "properties": {
          "value": {
            "type": "object"
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      },
      "PredicateResponse": {
        "type": "object",
        "required": [
          "value",
          "result"
        ],
        "properties": {
          "result": {
            "type": "boolean"
          },
          "value": {
            "type": "object"
          }
        }
      }
    }
  }
}